+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 
 ### FEATURES
 
+- [\#665](https://github.com/cosmos/evm/pull/665) Add EvmCodec address codec implementation
 - [\#346](https://github.com/cosmos/evm/pull/346) Add eth_createAccessList method and implementation
 - [\#337](https://github.com/cosmos/evm/pull/337) Support state overrides in eth_call.
 - [\#502](https://github.com/cosmos/evm/pull/502) Add block time in derived logs.